+++ .vscode/launch.json
@@ -68,7 +68,7 @@
       "args": [
         // "--extensionDevelopmentPath=${workspaceRoot}/e2e-tests/extension",
         "--extensionDevelopmentPath=${workspaceRoot}",
-        "--extensionTestsPath=${workspaceRoot}/e2e/out/index",
+        "--extensionTestsPath=${workspaceRoot}/e2e/out/runners/index",
         "${workspaceRoot}/e2e/projects/test-workspace.code-workspace"
       ],
       "env": {
@@ -85,7 +85,7 @@
       "runtimeExecutable": "${execPath}",
       "args": [
         "--extensionDevelopmentPath=${workspaceRoot}",
-        "--extensionTestsPath=${workspaceRoot}/e2e/out/dbt_ft",
+        "--extensionTestsPath=${workspaceRoot}/e2e/out/runners/dbt_ft",
         "${workspaceRoot}/../analytics/dbt_ft_prod"
       ],
       "env": {
